requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp_socks==0.10.1
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncache==0.3.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==44.0.3
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.0.1
faiss-cpu==1.11.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-google==0.1.1
langchain-google-genai==2.1.4
langchain-groq==0.3.2
langchain-ollama==0.3.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.42
litellm==1.69.1
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mdclense==0.1.2
mdurl==0.1.2
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
ollama==0.4.8
openai==1.75.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsera==0.2.5
pillow==10.4.0
playwright==1.52.0
playwright-stealth==1.0.6
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.0.0
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
readability==0.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
selenium==4.32.0
setuptools==80.3.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
tf-playwright-stealth==1.1.2
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
